# docs/openapi-roadmap.yaml
openapi: 3.0.3
info:
  title: SpriteForge Asset Production API
  version: 1.0.0
paths:
...
  /api/creative-documents/{document_id}:
    get:
      summary: Load a creative document revision
      responses: {'200': {description: Creative document}, '404': {description: Not found}}
    put:
      summary: Commit a document revision using optimistic concurrency
      responses: {'200': {description: Revision committed}, '409': {description: Revision conflict}}
...
  /api/studio/assets/{asset_id}/review:
    post:
      summary: Transition an asset revision through review
      responses: {'200': {description: Review state updated}, '403': {description: Insufficient role}}
...
  /api/farm/route:
    post:
      summary: Route work by capability, VRAM, quota, and current load
      responses: {'200': {description: Worker lease}, '409': {description: No eligible worker}}
...